started removing redundant code
